+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 ## 1.1.0
 
 - Added Security util for url signing.
+- Added support for custom domain and worker url
+- Added [upload](./README.md#upload) method.
 - Updated API endpoints